Add new layout cul-file-no-text
For now, left the original layout, file-no-text, on which
cul-file-no-text is based, alone. Basically, we just want to
rename this original layout, but since the Fugitive Slaves uses the
originally layout, both have to co-exist until the exhibition
has been updated to use the one with the new name (includes new
name for the subdir containing the layout files).