Fix file (image) alignment issue if original smaller than width
While testing, noticed that a certain image, with the File width
option set to "Full width", was always left aligned, regardless of the
File position setting. Turns out that the original image is smaller
than the width of a parent <div> which controls the width on the screen
(specified by the File width option). Trying different options
via inspect in Chrome, came up with the CSS changes associated with
this commit - basically, use text-align on a child <div> of the
aforementioned <div>
This bug was not noticed before because most images have a higher
resolution, and therefore fill completely the parent <div> which
controls the width.