Disable the cul-one-file-with-text and cul-gallery-no-text layouts
It was decided that these new block layouts may be confusing to
users. Code implementing the layouts was not removed or commented
out. Instead, the entries for the layouts in questions were commented
out of the exhibit_layouts filter.
Note that the leftover code will not pose a problem since it is in
subdirs specific to the layouts in question, and this code will
not be exercised while the layouts are disabled.
To enable the layouts, just remove the comments around the code
populating the entry in the exhibit_layouts filter.